Move task iterator to dedicated file
